requirements.txt
langchain==0.2.6
langchain-community==0.2.6
langchain-huggingface==0.0.3
faiss-cpu==1.8.0
sentence-transformers==2.7.0
llama-cpp-python==0.2.79
tqdm==4.66.4
requests==2.32.3
huggingface-hub==0.23.2
numpy==1.26.4
docx2txt==0.8
python-docx==1.1.2
pypdf==4.2.0
